+++ CHANGELOG.md
@@ -26,6 +26,8 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 **Next release**
 
 * Fixed all problems reported through the debug info
+* Added usage data white/black list help text
+* Updated Dutch translation
 * Updated French translation
 * Updated German translation
 * Updated Indonesian translation